djangoApp/requirement.txt
Django
djangorestframework
mysqlclient
djangorestframework-simplejwt
pillow
numpy
cv2